Move Ollama embedding models to OllamaModels.kt
